+++ oresoresores-1.18.2/common/src/main/java/ooo/OresOresOres.java
@@ -8,7 +8,7 @@ import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.Block;
-import ooo.block.Killium;
+import ooo.block.*;
 
 public class OresOresOres {
     public static final String MOD_ID = "ooo";
@@ -17,6 +17,12 @@ public class OresOresOres {
     public static final CreativeModeTab TAB = CreativeTabRegistry.create(new ResourceLocation(MOD_ID, "ores"), () -> new ItemStack(Killium.ITEM.get()));
 
     public static void init() {
+        Exordium.BLOCK = OresOresOres.BLOCKS.register("exordium", Exordium::new);
+        Exordium.ITEM = OresOresOres.ITEMS.register("exordium", Exordium.ExordiumItem::new);
+        RawExordium.ITEM = OresOresOres.ITEMS.register("raw_exordium", RawExordium::new);
+        ExordiumIngot.ITEM = OresOresOres.ITEMS.register("exordium_ingot", ExordiumIngot::new);
+        ExordiumBlock.BLOCK = OresOresOres.BLOCKS.register("exordium_block", ExordiumBlock::new);
+        ExordiumBlock.ITEM = OresOresOres.ITEMS.register("exordium_block", ExordiumBlock.ExordiumBlockItem::new);
         Killium.BLOCK = OresOresOres.BLOCKS.register("killium", Killium::new);
         Killium.ITEM = OresOresOres.ITEMS.register("killium", Killium.KilliumItem::new);
         BLOCKS.register();