+++ oresoresores-1.18.2/common/src/main/java/ooo/block/ExordiumIngot.java
@@ -1,0 +1,13 @@
+package ooo.block;
+
+import dev.architectury.registry.registries.RegistrySupplier;
+import net.minecraft.world.item.Item;
+import ooo.OresOresOres;
+
+public class ExordiumIngot extends Item {
+    public static RegistrySupplier<Item> ITEM;
+
+    public ExordiumIngot() {
+        super(new Properties().tab(OresOresOres.TAB));
+    }
+}