+++ changelog.md
@@ -1,2 +1,6 @@
 ## Ores Ores Ores %version%
 - Added Killium
+- Added Exordium
+  - Added Raw Exordium
+  - Added Exordium Ingot
+  - Added Exordium Block